measure AlphaBeta performance
