+++ src/main/scala/ai/scala/fp/game/mcts/perf/MctsDebugPerf.scala
@@ -1,10 +1,8 @@
 package ai.scala.fp.game.mcts.perf
 
-import ai.scala.fp.game.mcts.perf.Dummy._
+import ai.scala.fp.game.mcts.perf.DummyMcts._
 
 object MctsDebugPerf {
-
-
   def main(args: Array[String]): Unit = {
     dummyRoot.steps(1000000)
   }
